+++ .github/workflows/test.yml
@@ -10,9 +10,6 @@ jobs:
     runs-on: ubuntu-latest
     timeout-minutes: 10
 
-    env:
-      PYTHONPATH: ${{ github.workspace }}
-
     steps:
       - name: Checkout repo
         uses: actions/checkout@v2
