--- taxi/tests/test_views.py
@@ -1,24 +1,0 @@
-from django.test import TestCase
-from django.urls import reverse
-from django.contrib.auth import get_user_model
-from taxi.models import Manufacturer
-
-
-class ViewTests(TestCase):
-    def setUp(self):
-        self.user = get_user_model().objects.create_user(
-            username="testuser", password="testpass"
-        )
-        self.client.login(username="testuser", password="testpass")
-        Manufacturer.objects.create(name="Ford", country="USA")
-
-    def test_index_view(self):
-        response = self.client.get(reverse("taxi:index"))
-        self.assertEqual(response.status_code, 200)
-        self.assertTemplateUsed(response, "taxi/index.html")
-        self.assertIn("num_drivers", response.context)
-
-    def test_manufacturer_list_view(self):
-        response = self.client.get(reverse("taxi:manufacturer-list"))
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, "Ford")
